backend/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
python-dotenv==1.0.1
httpx==0.27.0
PyJWT[crypto]==2.9.0
redis==5.0.8
scikit-learn==1.4.2
pandas==2.2.2
numpy==1.26.4
openai==1.30.1
azure-identity==1.18.0
azure-keyvault-secrets==4.9.0
azure-ai-formrecognizer==3.3.3
azure-cognitiveservices-vision-face==0.6.0
pydantic==2.7.1
python-multipart==0.0.9
